Allowed lines to go 'past' a vertex to the edge
When drawing the ray, it will in addition to trying to draw the line from the source to that point, also try to draw from that point to the corresponding point on the border if the initial line was drawn.

Next step for this is for when the line past the vertex to the border intersects another edge.  In this case, it should draw the line from the point to the first edge hit (currently it just stops at the point).