It can shade!  It can shade!
Added a list that orders the points in which they should be shaded.  Points are added to that list whenever they have their corresponding lines drawn.  Implemented functions to compare which point should come before others.  I notice that it can sometimes jump around with shading as you move the mouse around, but I haven't been able to get it to stop on an incorrect shading ever.